Fix regression where failed join blocks daemon
server/handlers.go ConnectHandler() was failing to remove a key state
file on failure to connect. Daemon therefore fails to take control back
which results in user never being able to try to  connect to an AP
again. Fix is to remove the state file on failure to connect, thus
handing control back to daemon.